+++ projects/batfish/src/test/java/org/batfish/bddreachability/BDDMultipathInonsistencyWithLoop.java
@@ -1,0 +1,63 @@
+package org.batfish.bddreachability;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.hasSize;
+
+import com.google.common.collect.ImmutableSet;
+import java.io.IOException;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.SortedMap;
+import org.batfish.common.util.TracePruner;
+import org.batfish.datamodel.Configuration;
+import org.batfish.datamodel.Flow;
+import org.batfish.datamodel.FlowDisposition;
+import org.batfish.datamodel.acl.AclLineMatchExprs;
+import org.batfish.datamodel.flow.Trace;
+import org.batfish.main.Batfish;
+import org.batfish.main.BatfishTestUtils;
+import org.batfish.question.multipath.MultipathConsistencyParameters;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+
+public class BDDMultipathInonsistencyWithLoop {
+
+  @Rule public TemporaryFolder temp = new TemporaryFolder();
+
+  @Test
+  public void testLoopInMultipathConsistency() throws IOException {
+    SortedMap<String, Configuration> configurations =
+        new TestNetworkWithMPIandLoop()._configurations;
+    Batfish batfish = BatfishTestUtils.getBatfish(configurations, temp);
+    batfish.computeDataPlane(false);
+
+    MultipathConsistencyParameters multipathConsistencyParameters =
+        new MultipathConsistencyParameters(
+            AclLineMatchExprs.TRUE,
+            batfish.getAllSourcesInferFromLocationIpSpaceAssignment(),
+            configurations.keySet(),
+            ImmutableSet.of(),
+            TracePruner.DEFAULT_MAX_TRACES,
+            ImmutableSet.of());
+
+    Set<Flow> flows = batfish.bddMultipathConsistency(multipathConsistencyParameters);
+
+    // Flows starting from configuration 1 and configuration 2 both should result in inconsistent
+    // dispositions
+    assertThat(flows, hasSize(2));
+
+    Map<Flow, List<Trace>> flowTraces = batfish.buildFlows(flows, false);
+
+    // each flow should give out a list of traces with both the inconsistent dispositions
+    for (List<Trace> tracePerFlow : flowTraces.values()) {
+      Set<FlowDisposition> flowDispositions =
+          tracePerFlow.stream().map(Trace::getDisposition).collect(ImmutableSet.toImmutableSet());
+
+      assertThat(
+          flowDispositions, containsInAnyOrder(FlowDisposition.ACCEPTED, FlowDisposition.LOOP));
+    }
+  }
+}
